requirements.txt
neo4j
spacy 
chromadb 
llama-index 
streamlit 
fastapi 
uvicorn 
pandas 
numpy
python-dotenv
sentence-transformers 